+++ staramr/results/AMRDetectionSummary.py
@@ -16,7 +16,7 @@ class AMRDetectionSummary:
     SEPARATOR = ','
     FLOAT_DECIMALS = 2
 
-    def __init__(self, files, resfinder_dataframe: DataFrame, pointfinder_dataframe=None,
+    def __init__(self, files, resfinder_dataframe: DataFrame, quality_module_dataframe: DataFrame,pointfinder_dataframe=None,
                  plasmidfinder_dataframe=None, mlst_dataframe=None) -> None:
         """
         Constructs an object for summarizing AMR detection results.
@@ -28,21 +28,21 @@ class AMRDetectionSummary:
         self._resfinder_dataframe = resfinder_dataframe
         self._plasmidfinder_dataframe = plasmidfinder_dataframe
         self._mlst_dataframe = mlst_dataframe
-
         if pointfinder_dataframe is not None:
             self._has_pointfinder = True
             self._pointfinder_dataframe = pointfinder_dataframe
         else:
             self._has_pointfinder = False
+        self._quality_module_dataframe=quality_module_dataframe
 
     def _compile_results(self, resistance_frame: DataFrame) -> DataFrame:
         df_summary = resistance_frame.sort_values(by=['Gene']).groupby(['Isolate ID']).aggregate(
-            lambda x: {'Gene': (self.SEPARATOR + ' ').join(x['Gene'])})
+            lambda x: {'Gene': (self.SEPARATOR + ' ').join(x.get('Gene'))})
         return df_summary[['Gene']]
 
     def _compile_plasmids(self, plasmid_frame: DataFrame) -> DataFrame:
         ds_summary = plasmid_frame.sort_values(by=['Gene']).groupby(['Isolate ID']).aggregate(
-            lambda x: {'Gene': (self.SEPARATOR + ' ').join(x['Gene'])})
+            lambda x: {'Gene': (self.SEPARATOR + ' ').join(x.get('Gene'))})
 
         ds_frame = ds_summary[['Gene']]
 
@@ -156,6 +156,13 @@ class AMRDetectionSummary:
         if mlst_frame is not None:
             mlst_merging_frame = mlst_frame[['Scheme', 'Sequence Type']]
             resistance_frame = resistance_frame.merge(mlst_merging_frame, on='Isolate ID', how='left')
+
+
+
+        resistance_frame = resistance_frame.merge(self._quality_module_dataframe, on='Isolate ID', how='left')
+
+        #Rearranges the resistance frame so that the Quality Module column comes directly after Isolate ID
+        resistance_frame = resistance_frame[['Quality Module'] + [col for col in resistance_frame if col not in ['Quality Module']]] 
 
         return resistance_frame.sort_index()